rework stack parser
