--- posthog-node/src/extensions/error-tracking/stack-parser.edge.ts
@@ -1,3 +1,0 @@
-import { StackParser } from './types'
-
-export const defaultStackParser: StackParser = () => []
