+++ posthog-node/src/extensions/error-tracking/types.ts
@@ -42,6 +42,8 @@ export interface Mechanism {
   synthetic?: boolean
 }
 
+export type GetModuleFn = (filename: string | undefined) => string | undefined
+
 export type StackParser = (stack: string, skipFirstLines?: number) => StackFrame[]
 export type StackLineParserFn = (line: string) => StackFrame | undefined
 export type StackLineParser = [number, StackLineParserFn]
